refactor: move policy stack (policy, confidence, edm, provenance) to src/enforcement
No behavior change; src/proxy retains only transport (jsonrpc, proxy-core,
installer, audit-log). Full suite at the 2393-test baseline.
